+++ changelog.md
@@ -1,9 +1,15 @@
-### 0.0.3 "initial public release" 2022/03/16
+### 0.0.4 "multi contents" 2022/03/17
+
+ * Toaster.show(*content) can take multiple objects
+ * MBox.show(*content) can take multiple objects
+ * MBox.confirm(content, ok, ko=None,txtok="OK",txtko="Cancel") (ok/ko are callbacks)
+
+### 0.0.3 "classEnsure ensure" 2022/03/16
 
  * __version__ is uptodate to toml version
  * classEnsure() is a little bit better (not appending continously)
 
-### 0.0.2 "initial public release" 2022/03/16
+### 0.0.2 "right imports" 2022/03/16
 
  * just the package imports